YT-27710: Add RPC to TransactionPinger
* Changelog entry
  Type: fix
  Component: cpp-sdk

Allow TransactionPinger to ping via RPC if is being used in RPC Client.
commit_hash:9b38ce40b5667f107d6f23e40bf07dcf80ac11c7
