+++ yt/cpp/mapreduce/client/transaction_pinger.cpp
@@ -2,30 +2,15 @@
 
 #include "transaction.h"
 
-#include <yt/cpp/mapreduce/interface/config.h>
 #include <yt/cpp/mapreduce/interface/error_codes.h>
 #include <yt/cpp/mapreduce/interface/logging/yt_log.h>
-#include <yt/cpp/mapreduce/interface/tvm.h>
-
-#include <yt/cpp/mapreduce/common/wait_proxy.h>
-#include <yt/cpp/mapreduce/common/retry_lib.h>
 
 #include <yt/cpp/mapreduce/http/requests.h>
 #include <yt/cpp/mapreduce/http/retry_request.h>
 
 #include <yt/yt/core/concurrency/periodic_executor.h>
-#include <yt/yt/core/concurrency/poller.h>
 #include <yt/yt/core/concurrency/scheduler_api.h>
-#include <yt/yt/core/concurrency/thread_pool_poller.h>
 #include <yt/yt/core/concurrency/thread_pool.h>
-
-#include <yt/yt/core/http/client.h>
-#include <yt/yt/core/http/http.h>
-
-#include <yt/yt/core/https/client.h>
-#include <yt/yt/core/https/config.h>
-
-#include <library/cpp/yson/node/node_io.h>
 
 #include <library/cpp/yt/threading/spin_lock.h>
 #include <library/cpp/yt/assert/assert.h>
@@ -37,85 +22,13 @@ namespace NYT {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-namespace {
-
-////////////////////////////////////////////////////////////////////////////////
-
-void CheckError(const TString& requestId, NHttp::IResponsePtr response)
-{
-    if (const auto* ytError = response->GetHeaders()->Find("X-YT-Error")) {
-        TYtError error;
-        error.ParseFrom(*ytError);
-
-        TErrorResponse errorResponse(std::move(error), requestId);
-        if (errorResponse.IsOk()) {
-            return;
-        }
-
-        YT_LOG_ERROR("RSP %v - HTTP %v - %v",
-            requestId,
-            response->GetStatusCode(),
-            errorResponse.AsStrBuf());
-
-        ythrow errorResponse;
-    }
-}
-
-void PingTx(NHttp::IClientPtr httpClient, const TPingableTransaction& tx)
-{
-    const auto context = tx.GetContext();
-    auto url = TString::Join(context.UseTLS ? "https://" : "http://", context.ServerName, "/api/", context.Config->ApiVersion, "/ping_tx");
-    auto headers = New<NHttp::THeaders>();
-    auto requestId = CreateGuidAsString();
-
-    headers->Add("Host", url);
-    headers->Add("User-Agent", TProcessState::Get()->ClientVersion);
-
-    if (const auto& serviceTicketAuth = context.ServiceTicketAuth) {
-        const auto serviceTicket = serviceTicketAuth->Ptr->IssueServiceTicket();
-        headers->Add("X-Ya-Service-Ticket", serviceTicket);
-    } else if (const auto& token = context.Token; !token.empty()) {
-        headers->Add("Authorization", "OAuth " + token);
-    }
-
-    headers->Add("Transfer-Encoding", "chunked");
-    headers->Add("X-YT-Correlation-Id", requestId);
-    headers->Add("X-YT-Header-Format", "<format=text>yson");
-    headers->Add("Content-Encoding", "identity");
-    headers->Add("Accept-Encoding", "identity");
-
-    TNode node;
-    node["transaction_id"] = GetGuidAsString(tx.GetId());
-    auto strParams = NodeToYsonString(node);
-
-    YT_LOG_DEBUG("REQ %v - sending request (HostName: %v; Method POST %v; X-YT-Parameters (sent in body): %v)",
-        requestId,
-        context.ServerName,
-        url,
-        strParams
-    );
-
-    auto response = NConcurrency::WaitFor(httpClient->Post(url, TSharedRef::FromString(strParams), headers)).ValueOrThrow();
-    CheckError(requestId, response);
-
-    YT_LOG_DEBUG("RSP %v - received response %v bytes. (%v)",
-            requestId,
-            response->ReadAll().size(),
-            strParams);
-}
-
-} // namespace
-
-////////////////////////////////////////////////////////////////////////////////
-
 class TSharedTransactionPinger
     : public ITransactionPinger
 {
 public:
-    TSharedTransactionPinger(NHttp::IClientPtr httpClient, int poolThreadCount)
+    explicit TSharedTransactionPinger(int poolThreadCount)
         : PingerPool_(NConcurrency::CreateThreadPool(
             poolThreadCount, "tx_pinger_pool"))
-        , HttpClient_(std::move(httpClient))
     { }
 
     ~TSharedTransactionPinger() override
@@ -181,7 +94,7 @@ private:
                            NConcurrency::TPeriodicExecutorPtr periodic)
     {
         try {
-            PingTx(HttpClient_, pingableTx);
+            pingableTx.Ping();
         } catch (const TErrorResponse& e) {
             /// NB: No logging here, CheckError() already logged TErrorResponse.
             if (e.GetError().ContainsErrorCode(NYT::NClusterErrorCodes::NTransactionClient::NoSuchTransaction)) {
@@ -202,32 +115,15 @@ private:
     THashMap<TTransactionId, std::shared_ptr<NConcurrency::TPeriodicExecutorPtr>> Transactions_;
 
     NConcurrency::IThreadPoolPtr PingerPool_;
-    NHttp::IClientPtr HttpClient_;
 };
 
 ////////////////////////////////////////////////////////////////////////////////
 
-ITransactionPingerPtr CreateTransactionPinger(const TConfigPtr& config, bool useTLS)
+ITransactionPingerPtr CreateTransactionPinger(const TConfigPtr& config)
 {
     YT_LOG_DEBUG("Using async transaction pinger");
-    auto httpPoller = NConcurrency::CreateThreadPoolPoller(
-        config->AsyncHttpClientThreads,
-        "tx_http_client_poller");
-    NHttp::IClientPtr httpClient;
 
-    if (useTLS) {
-        auto httpsClientConfig = NYT::New<NHttps::TClientConfig>();
-        httpsClientConfig->MaxIdleConnections = 16;
-        httpClient = NHttps::CreateClient(std::move(httpsClientConfig), std::move(httpPoller));
-    } else {
-        auto httpClientConfig = NYT::New<NHttp::TClientConfig>();
-        httpClientConfig->MaxIdleConnections = 16;
-        httpClient = NHttp::CreateClient(std::move(httpClientConfig), std::move(httpPoller));
-    }
-
-    return MakeIntrusive<TSharedTransactionPinger>(
-        std::move(httpClient),
-        config->AsyncTxPingerPoolThreads);
+    return MakeIntrusive<TSharedTransactionPinger>(config->AsyncTxPingerPoolThreads);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
